refactor: reorganize register implementation by architecture
- Move aarch64 specific register code to dedicated module
- Add x86_64 register implementation in separate module
- Update main register module to conditionally import based on target architecture
